check if hit misses in raycast
Without checking if hit returns a valid result lib show text will remain on the screen until you use another seed and try again.  This will hide the UI and allow you to return to the top of the event cleaner. rayCastDistance also seems to perform better as 10.0 from testing.